+++ example.env
@@ -40,8 +40,8 @@ NOTIFICATION_IMPLEMENTATION="memory"
 # WEBPUSH_VAPID_PUBLIC_KEY=""
 # WEBPUSH_VAPID_PRIVATE_KEY=""
 
-# SANITIZER (sanitize-html)
-SANITIZER_IMPLEMENTATION="sanitize-html"
+# SANITIZATION (sanitize-html)
+SANITIZATION_IMPLEMENTATION="sanitize-html"
 
 # VALIDATION (zod)
 VALIDATION_IMPLEMENTATION="zod"
