+++ src/integrations/sanitization/definitions/interfaces.ts
