+++ src/integrations/sanitization/errors/SanitizerError.ts
